+++ tf/compose/__init__.py
@@ -2,5 +2,7 @@
 .. include:: ../../docs/compose/top.md
 """
 
+
 from .modify import modify
 from .combine import combine
+from .nodemaps import Versions